requirements.txt
## The following requirements were added by pip --freeze:
Django==1.7
Pillow==2.5.3
argparse==1.2.1
coverage==3.7.1
django-debug-toolbar==1.2.1
factory-boy==2.4.1
mock==1.0.1
pep8==1.5.7
psycopg2==2.5.4
pyflakes==0.8.1
requests==2.4.3
selenium==2.43.0
sqlparse==0.1.11
wsgiref==0.1.2
